LPS-197423 Move LayoutData types to it's own directory
