+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/actions/updateColSize.ts
@@ -5,7 +5,7 @@
 
 import {UPDATE_COL_SIZE} from './types';
 
-import type {LayoutData} from '../../types/LayoutData';
+import type {LayoutData} from '../../types/layout_data/LayoutData';
 
 export default function updateColSize({
 	layoutData,
