+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/actions/updateCollectionDisplayCollection.ts
@@ -5,7 +5,7 @@
 
 import {UPDATE_COLLECTION_DISPLAY_COLLECTION} from './types';
 
-import type {LayoutData} from '../../types/LayoutData';
+import type {LayoutData} from '../../types/layout_data/LayoutData';
 import type {FragmentEntryLink} from './addFragmentEntryLinks';
 import type {PageContent} from './addItem';
