+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/actions/updateFormItemConfig.ts
@@ -5,7 +5,10 @@
 
 import {UPDATE_FORM_ITEM_CONFIG} from './types';
 
-import type {DeletedLayoutDataItem, LayoutData} from '../../types/LayoutData';
+import type {
+	DeletedLayoutDataItem,
+	LayoutData,
+} from '../../types/layout_data/LayoutData';
 import type {FragmentEntryLinkMap} from './addFragmentEntryLinks';
 
 export default function updateFormItemConfig({
