+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/actions/updateRule.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {LayoutData} from '../../types/LayoutData';
+import {LayoutData} from '../../types/layout_data/LayoutData';
 import {UPDATE_RULE} from './types';
 
 type Props = {
