+++ modules/apps/layout/layout-content-page-editor-web/types/src/main/resources/META-INF/resources/page_editor/app/actions/updateFormItemConfig.d.ts
@@ -3,7 +3,10 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import type {DeletedLayoutDataItem, LayoutData} from '../../types/LayoutData';
+import type {
+	DeletedLayoutDataItem,
+	LayoutData,
+} from '../../types/layout_data/LayoutData';
 import type {FragmentEntryLinkMap} from './addFragmentEntryLinks';
 export default function updateFormItemConfig({
 	addedFragmentEntryLinks,
